Issue 78: Move domain and application user elements All now live in the user component instead of the UserBundle
