+++ src/Khatovar/Bundle/UserBundle/Controller/Admin/UpdateController.php
@@ -23,10 +23,10 @@ declare(strict_types=1);
 
 namespace Khatovar\Bundle\UserBundle\Controller\Admin;
 
-use Khatovar\Bundle\UserBundle\Event\UserEvents;
 use Khatovar\Bundle\UserBundle\Form\Factory\UserFormFactory;
 use Khatovar\Bundle\UserBundle\Form\Type\UserType;
-use Khatovar\Bundle\UserBundle\Query\GetUserFromDatabase;
+use Khatovar\Component\User\Application\Query\GetUser;
+use Khatovar\Component\User\Domain\Event\UserEvents;
 use Khatovar\Component\User\Domain\Exception\UserDoesNotExist;
 use Symfony\Bridge\Doctrine\RegistryInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -48,7 +48,7 @@ use Twig\Environment;
  */
 class UpdateController
 {
-    /** @var GetUserFromDatabase */
+    /** @var GetUser */
     private $getUser;
 
     /** @var UserFormFactory */
@@ -73,7 +73,7 @@ class UpdateController
     private $twig;
 
     /**
-     * @param GetUserFromDatabase      $getUser
+     * @param GetUser                  $getUser
      * @param UserFormFactory          $userFormFactory
      * @param EventDispatcherInterface $eventDispatcher
      * @param RegistryInterface        $doctrine
@@ -83,7 +83,7 @@ class UpdateController
      * @param Environment              $twig
      */
     public function __construct(
-        GetUserFromDatabase $getUser,
+        GetUser $getUser,
         UserFormFactory $userFormFactory,
         EventDispatcherInterface $eventDispatcher,
         RegistryInterface $doctrine,
