+++ src/Khatovar/Component/User/Application/Query/GetUser.php
@@ -21,10 +21,10 @@ declare(strict_types=1);
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Khatovar\Bundle\UserBundle\Query;
+namespace Khatovar\Component\User\Application\Query;
 
-use Khatovar\Bundle\UserBundle\Entity\UserInterface;
 use Khatovar\Component\User\Domain\Exception\UserDoesNotExist;
+use Khatovar\Component\User\Domain\Model\UserInterface;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
 /**
