+++ src/Khatovar/Component/User/Domain/Repository/UserRepositoryInterface.php
@@ -5,15 +5,25 @@ declare(strict_types=1);
 /*
  * This file is part of KhatovarWeb.
  *
- * Copyright (c) 2016 Damien Carcel <damien.carcel@gmail.com>
+ * Copyright (c) 2018 Damien Carcel <damien.carcel@gmail.com>
  *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-namespace Khatovar\Bundle\UserBundle\Entity\Repository;
+namespace Khatovar\Component\User\Domain\Repository;
 
-use Khatovar\Bundle\UserBundle\Entity\UserInterface;
+use Khatovar\Component\User\Domain\Model\UserInterface;
 
 /**
  * User repository interface.
