+++ tests/spec/Khatovar/Bundle/UserBundle/Handler/UserStatusHandlerSpec.php
@@ -11,8 +11,8 @@
 
 namespace spec\Khatovar\Bundle\UserBundle\Handler;
 
-use Khatovar\Bundle\UserBundle\Entity\UserInterface;
-use Khatovar\Bundle\UserBundle\Event\UserEvents;
+use Khatovar\Component\User\Domain\Model\UserInterface;
+use Khatovar\Component\User\Domain\Event\UserEvents;
 use Khatovar\Bundle\UserBundle\Handler\UserStatusHandler;
 use Khatovar\Bundle\UserBundle\Handler\UserStatusHandlerInterface;
 use Doctrine\ORM\EntityManagerInterface;
