+++ tests/spec/Khatovar/Bundle/UserBundle/Manager/RolesManagerSpec.php
@@ -11,9 +11,9 @@
 
 namespace spec\Khatovar\Bundle\UserBundle\Manager;
 
-use Khatovar\Bundle\UserBundle\Entity\UserInterface;
+use Khatovar\Component\User\Domain\Model\UserInterface;
 use Khatovar\Bundle\UserBundle\Manager\RolesManager;
-use Khatovar\Bundle\UserBundle\Security\Core\Authentication\CurrentUser;
+use Khatovar\Component\User\Application\Query\CurrentUser;
 use PhpSpec\ObjectBehavior;
 
 /**
